Task Assistant/requirements.txt
google-generativeai
python-dotenv
schedule